FIX #109 - Refactor io folder
