+++ kedro_mlflow/framework/hooks/pipeline_hook.py
@@ -12,7 +12,7 @@ from kedro.versioning.journal import _git_sha
 from mlflow.models import infer_signature
 
 from kedro_mlflow.framework.context import get_mlflow_config
-from kedro_mlflow.io import MlflowMetricsDataSet
+from kedro_mlflow.io.metrics import MlflowMetricsDataSet
 from kedro_mlflow.mlflow import KedroPipelineModel
 from kedro_mlflow.pipeline.pipeline_ml import PipelineML
 from kedro_mlflow.utils import _parse_requirements
