+++ kedro_mlflow/io/artifacts/__init__.py
@@ -1,0 +1,1 @@
+from .mlflow_artifact_dataset import MlflowArtifactDataSet
