+++ kedro_mlflow/io/metrics/__init__.py
@@ -1,2 +1,1 @@
-from .mlflow_dataset import MlflowArtifactDataSet
 from .mlflow_metrics_dataset import MlflowMetricsDataSet
